Tried adding Firebase sign in
