+++ android/app/google-services.json
@@ -1,0 +1,47 @@
+{
+  "project_info": {
+    "project_number": "233330185060",
+    "project_id": "brain-exercise-initiative",
+    "storage_bucket": "brain-exercise-initiative.appspot.com"
+  },
+  "client": [
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:233330185060:android:0dd71b83572d1de7eba071",
+        "android_client_info": {
+          "package_name": "com.BrainExerciseInitiative.BrainExercise"
+        }
+      },
+      "oauth_client": [
+        {
+          "client_id": "233330185060-j6qfhlkgi3f1f952cbq6h0asrhg3lhsi.apps.googleusercontent.com",
+          "client_type": 1,
+          "android_info": {
+            "package_name": "com.BrainExerciseInitiative.BrainExercise",
+            "certificate_hash": "4bb0b34e459b321f6a1d51dd46c6566578bf2910"
+          }
+        },
+        {
+          "client_id": "233330185060-882ehus1tj0beov35ciaiggnsvteposn.apps.googleusercontent.com",
+          "client_type": 3
+        }
+      ],
+      "api_key": [
+        {
+          "current_key": "AIzaSyAgrRCLsVflTkAx4QDXneWbrcO37hJ_5ic"
+        }
+      ],
+      "services": {
+        "appinvite_service": {
+          "other_platform_oauth_client": [
+            {
+              "client_id": "233330185060-0afttv4rctfe4ijscqq9npbd2vtag1s4.apps.googleusercontent.com",
+              "client_type": 3
+            }
+          ]
+        }
+      }
+    }
+  ],
+  "configuration_version": "1"
+}
