+++ package.json
@@ -22,6 +22,7 @@
     "@react-native-async-storage/async-storage": "1.18.2",
     "@react-native-community/datetimepicker": "7.2.0",
     "@react-native-community/slider": "4.4.2",
+    "@react-native-google-signin/google-signin": "^10.0.1",
     "@react-native-masked-view/masked-view": "^0.2.9",
     "@react-navigation/native": "^6.1.7",
     "@react-navigation/native-stack": "^6.9.13",
@@ -35,6 +36,7 @@
     "expo": "^49.0.0",
     "expo-auth-session": "~5.0.2",
     "expo-av": "~13.4.1",
+    "expo-build-properties": "^0.8.3",
     "expo-constants": "~14.4.2",
     "expo-dev-client": "~2.4.8",
     "expo-notifications": "~0.20.1",
